refactor: 중복 코드 interface의 default method로 통합
