+++ src/main/java/domain/piece/Piece.java
@@ -3,9 +3,23 @@ package domain.piece;
 import domain.board.Board;
 import domain.board.Node;
 
+import java.util.List;
+
 public interface Piece {
 
-    boolean canMove(Node source, Node destination, Board board);
+    default boolean canMove(Node source, Node destination, Board board) {
+        return isDifferentNode(source, destination) && containsInCandidates(source, destination, board);
+    }
+
+    private boolean isDifferentNode(Node source, Node destination) {
+        return !destination.isSameNode(source);
+    }
+
+    private boolean containsInCandidates(Node source, Node destination, Board board) {
+        return findMovableNodes(source, board).contains(destination);
+    }
+
+    List<Node> findMovableNodes(Node source, Board board);
 
     PieceType type();
