+++ src/main/java/domain/piece/Sang.java
@@ -56,19 +56,7 @@ public class Sang implements Piece {
     }
 
     @Override
-    public boolean canMove(Node source, Node destination, Board board) {
-        return isDifferentNode(source, destination) && containsInCandidates(source, destination, board);
-    }
-
-    private boolean isDifferentNode(Node source, Node destination) {
-        return !destination.isSameNode(source);
-    }
-
-    private boolean containsInCandidates(Node source, Node destination, Board board) {
-        return findMovableNodes(source, board).contains(destination);
-    }
-
-    private List<Node> findMovableNodes(Node sourceNode, Board board) {
+    public List<Node> findMovableNodes(Node sourceNode, Board board) {
         List<Node> candidates = new ArrayList<>();
         for (PiecePath piecePath : PIECE_PATHS) {
             checkObstaclesAndAddCandidate(sourceNode,
